+++ src/option_pricing/exotic/asian.cpp
@@ -31,7 +31,7 @@ AsianOptionGeometric::AsianOptionGeometric(PayOff* _pay_off) : AsianOption(_pay_
 double AsianOptionGeometric::pay_off_price(const std::vector<double>& spot_prices) const {
     unsigned num_times = spot_prices.size();
     double log_sum = 0.0;
-    for (int i = 0; i < spot_prices.size(); i++) {
+    for (size_t i = 0; i < spot_prices.size(); i++) {
         log_sum += log(spot_prices[i]);
     }
     double geom_mean = exp(log_sum / static_cast<double>(num_times));
